requirements.txt
# Core deep learning framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers and vision-language models
transformers>=4.30.0
open-clip-torch>=2.20.0

# Image processing
Pillow>=10.0.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
networkx>=3.0

# Configuration
PyYAML>=6.0

# Progress bars and utilities
tqdm>=4.65.0

# Testing
pytest>=7.0.0

# Optional: for CUDA support (PyTorch will install appropriate CUDA libraries)
# Uncomment if you need specific CUDA versions
# nvidia-cuda-runtime-cu12>=12.0.0
# nvidia-cudnn-cu12>=9.0.0
